Renamed CASMonitorHttpClient to the more specific DataHubMonitorHttpClient. Moved the class to a datahub specific location.
